Rename namespace device to hardware
