+++ oneflow/core/device/cuda_util.cpp
@@ -17,8 +17,8 @@ limitations under the License.
 #include "oneflow/core/device/cuda_util.h"
 #include "oneflow/core/common/global.h"
 #include "oneflow/core/common/multi_client.h"
-#include "oneflow/core/device/node_device_descriptor_manager.h"
-#include "oneflow/core/device/cuda_device_descriptor.h"
+#include "oneflow/core/hardware/node_device_descriptor_manager.h"
+#include "oneflow/core/hardware/cuda_device_descriptor.h"
 #include "oneflow/core/rpc/include/global_process_ctx.h"
 #include "oneflow/core/job/env_global_objects_scope.h"
 #include "oneflow/core/job/lazy_mode.h"
@@ -129,11 +129,11 @@ namespace {
 
 std::function<cudaError_t(void**, size_t)> GetCudaMallocHostFn(int32_t dev) {
   auto default_fn = [](void** ptr, size_t size) { return cudaMallocHost(ptr, size); };
-  auto manager = Global<device::NodeDeviceDescriptorManager>::Get();
+  auto manager = Global<hardware::NodeDeviceDescriptorManager>::Get();
   if (manager == nullptr) { return default_fn; }
   auto node_desc = manager->GetLocalNodeDeviceDescriptor();
-  auto cuda_device = std::dynamic_pointer_cast<const device::CudaDeviceDescriptor>(
-      node_desc->GetDevice(device::kCudaDeviceDescriptorClassName, dev));
+  auto cuda_device = std::dynamic_pointer_cast<const hardware::CudaDeviceDescriptor>(
+      node_desc->GetDevice(hardware::kCudaDeviceDescriptorClassName, dev));
   if (!cuda_device) { return default_fn; }
   auto saved_affinity = node_desc->Topology()->GetMemoryAffinity();
   if (!saved_affinity) { return default_fn; }
