+++ oneflow/core/ep/cuda/cuda_stream.cpp
@@ -16,8 +16,8 @@ limitations under the License.
 #include "oneflow/core/ep/cuda/cuda_stream.h"
 #include "oneflow/core/job/global_for.h"
 #include "oneflow/core/job/resource_desc.h"
-#include "oneflow/core/device/node_device_descriptor_manager.h"
-#include "oneflow/core/device/cuda_device_descriptor.h"
+#include "oneflow/core/hardware/node_device_descriptor_manager.h"
+#include "oneflow/core/hardware/cuda_device_descriptor.h"
 #include "oneflow/core/ep/cuda/cuda_event.h"
 #include "oneflow/core/ep/cuda/cuda_device.h"
 
@@ -32,11 +32,11 @@ namespace {
 constexpr size_t kDefaultWorkspaceSize = 4 * 1024 * 1024;  // 4M
 
 void SetAffinityByDevice(int dev_id) {
-  auto node_device_desc_mgr = Global<device::NodeDeviceDescriptorManager>::Get();
+  auto node_device_desc_mgr = Global<hardware::NodeDeviceDescriptorManager>::Get();
   if (node_device_desc_mgr == nullptr) { return; }
   auto node_device_desc = node_device_desc_mgr->GetLocalNodeDeviceDescriptor();
-  auto cuda_device = std::dynamic_pointer_cast<const device::CudaDeviceDescriptor>(
-      node_device_desc->GetDevice(device::kCudaDeviceDescriptorClassName, dev_id));
+  auto cuda_device = std::dynamic_pointer_cast<const hardware::CudaDeviceDescriptor>(
+      node_device_desc->GetDevice(hardware::kCudaDeviceDescriptorClassName, dev_id));
   if (!cuda_device) { return; }
   node_device_desc->Topology()->SetCPUAffinityByPCIBusID(cuda_device->PCIBusID());
   node_device_desc->Topology()->SetMemoryAffinityByPCIBusID(cuda_device->PCIBusID());
