+++ oneflow/core/hardware/basic_device_descriptor_list.cpp
@@ -13,11 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/device/basic_device_descriptor_list.h"
+#include "oneflow/core/hardware/basic_device_descriptor_list.h"
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
 BasicDeviceDescriptorList::BasicDeviceDescriptorList(
     std::vector<std::shared_ptr<const DeviceDescriptor>> device_descriptor_list)
@@ -38,6 +38,6 @@ std::shared_ptr<const DeviceDescriptor> BasicDeviceDescriptorList::GetDevice(siz
   }
 }
 
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
