+++ oneflow/core/hardware/cuda_device_descriptor.h
@@ -13,10 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_DEVICE_CUDA_DEVICE_DESCRIPTOR_H_
-#define ONEFLOW_CORE_DEVICE_CUDA_DEVICE_DESCRIPTOR_H_
+#ifndef ONEFLOW_CORE_HARDWARE_CUDA_DEVICE_DESCRIPTOR_H_
+#define ONEFLOW_CORE_HARDWARE_CUDA_DEVICE_DESCRIPTOR_H_
 
-#include "oneflow/core/device/device_descriptor.h"
+#include "oneflow/core/hardware/device_descriptor.h"
+#include "oneflow/core/common/util.h"
 #include <string>
 #include <memory>
 
@@ -24,12 +25,13 @@ limitations under the License.
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
 constexpr char kCudaDeviceDescriptorClassName[] = "cuda";
 
 class CudaDeviceDescriptor : public DeviceDescriptor {
  public:
+  OF_DISALLOW_COPY_AND_MOVE(CudaDeviceDescriptor);
   ~CudaDeviceDescriptor() override;
 
   int32_t Ordinal() const;
@@ -52,10 +54,10 @@ class CudaDeviceDescriptor : public DeviceDescriptor {
   std::unique_ptr<Impl> impl_;
 };
 
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
 
 #endif  // WITH_CUDA
 
-#endif  // ONEFLOW_CORE_DEVICE_CUDA_DEVICE_DESCRIPTOR_H_
+#endif  // ONEFLOW_CORE_HARDWARE_CUDA_DEVICE_DESCRIPTOR_H_
