+++ oneflow/core/hardware/device_descriptor.h
@@ -13,21 +13,24 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+#ifndef ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_H_
+#define ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_H_
 
-#include "oneflow/core/device/topology_descriptor.h"
+#include "oneflow/core/common/util.h"
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
-void TopologyDescriptor::SetCPUAffinityByPCIBusID(const std::string& bus_id) const {
-  SetCPUAffinity(GetCPUAffinityByPCIBusID(bus_id));
-}
+class DeviceDescriptor {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(DeviceDescriptor);
+  DeviceDescriptor() = default;
+  virtual ~DeviceDescriptor() = default;
+};
 
-void TopologyDescriptor::SetMemoryAffinityByPCIBusID(const std::string& bus_id) const {
-  SetMemoryAffinity(GetMemoryAffinityByPCIBusID(bus_id));
-}
-
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
+
+#endif  // ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_H_
