+++ oneflow/core/hardware/device_descriptor_class.h
@@ -13,17 +13,19 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_DEVICE_DEVICE_DESCRIPTOR_CLASS_H_
-#define ONEFLOW_CORE_DEVICE_DEVICE_DESCRIPTOR_CLASS_H_
+#ifndef ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_CLASS_H_
+#define ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_CLASS_H_
 
-#include "oneflow/core/device/device_descriptor_list.h"
+#include "oneflow/core/hardware/device_descriptor_list.h"
+#include "oneflow/core/common/util.h"
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
 class DeviceDescriptorClass {
  public:
+  OF_DISALLOW_COPY_AND_MOVE(DeviceDescriptorClass);
   DeviceDescriptorClass() = default;
   virtual ~DeviceDescriptorClass() = default;
 
@@ -43,8 +45,8 @@ class DeviceDescriptorClass {
       const std::string& class_name);
 };
 
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_DEVICE_DEVICE_DESCRIPTOR_CLASS_H_
+#endif  // ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_CLASS_H_
