+++ oneflow/core/hardware/device_descriptor_list.h
@@ -13,21 +13,30 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_DEVICE_DEVICE_DESCRIPTOR_H_
-#define ONEFLOW_CORE_DEVICE_DEVICE_DESCRIPTOR_H_
+#ifndef ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_LIST_H_
+#define ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_LIST_H_
+
+#include "oneflow/core/hardware/device_descriptor.h"
+#include "oneflow/core/common/util.h"
+#include <cstdint>
+#include <memory>
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
-class DeviceDescriptor {
+class DeviceDescriptorList {
  public:
-  DeviceDescriptor() = default;
-  virtual ~DeviceDescriptor() = default;
+  OF_DISALLOW_COPY_AND_MOVE(DeviceDescriptorList);
+  DeviceDescriptorList() = default;
+  virtual ~DeviceDescriptorList() = default;
+
+  virtual size_t DeviceCount() const = 0;
+  virtual std::shared_ptr<const DeviceDescriptor> GetDevice(size_t ordinal) const = 0;
 };
 
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_DEVICE_DEVICE_DESCRIPTOR_H_
+#endif  // ONEFLOW_CORE_HARDWARE_DEVICE_DESCRIPTOR_LIST_H_
