+++ oneflow/core/hardware/net_ib_device_descriptor.h
@@ -13,10 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_DEVICE_NET_IB_DEVICE_DESCRIPTOR_H_
-#define ONEFLOW_CORE_DEVICE_NET_IB_DEVICE_DESCRIPTOR_H_
+#ifndef ONEFLOW_CORE_HARDWARE_NET_IB_DEVICE_DESCRIPTOR_H_
+#define ONEFLOW_CORE_HARDWARE_NET_IB_DEVICE_DESCRIPTOR_H_
 
-#include "oneflow/core/device/device_descriptor.h"
+#include "oneflow/core/hardware/device_descriptor.h"
+#include "oneflow/core/common/util.h"
 #include <string>
 #include <memory>
 
@@ -26,7 +27,7 @@ limitations under the License.
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
 constexpr char kNetIBDeviceDescriptorClassName[] = "net_ib";
 
@@ -38,6 +39,7 @@ enum NetIBDeviceDescriptorLinkLayer {
 
 class NetIBDeviceDescriptor : public DeviceDescriptor {
  public:
+  OF_DISALLOW_COPY_AND_MOVE(NetIBDeviceDescriptor);
   ~NetIBDeviceDescriptor() override;
 
   int32_t Ordinal() const;
@@ -58,10 +60,10 @@ class NetIBDeviceDescriptor : public DeviceDescriptor {
   std::unique_ptr<Impl> impl_;
 };
 
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
 
 #endif  // WITH_RDMA
 
-#endif  // ONEFLOW_CORE_DEVICE_NET_IB_DEVICE_DESCRIPTOR_H_
+#endif  // ONEFLOW_CORE_HARDWARE_NET_IB_DEVICE_DESCRIPTOR_H_
