+++ oneflow/core/hardware/node_device_descriptor_manager.h
@@ -13,17 +13,19 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_DEVICE_NODE_DEVICE_DESCRIPTOR_MANAGER_H_
-#define ONEFLOW_CORE_DEVICE_NODE_DEVICE_DESCRIPTOR_MANAGER_H_
+#ifndef ONEFLOW_CORE_HARDWARE_NODE_DEVICE_DESCRIPTOR_MANAGER_H_
+#define ONEFLOW_CORE_HARDWARE_NODE_DEVICE_DESCRIPTOR_MANAGER_H_
 
-#include "oneflow/core/device/node_device_descriptor.h"
+#include "oneflow/core/hardware/node_device_descriptor.h"
+#include "oneflow/core/common/util.h"
 
 namespace oneflow {
 
-namespace device {
+namespace hardware {
 
 class NodeDeviceDescriptorManager {
  public:
+  OF_DISALLOW_COPY_AND_MOVE(NodeDeviceDescriptorManager);
   NodeDeviceDescriptorManager();
   ~NodeDeviceDescriptorManager();
 
@@ -37,8 +39,8 @@ class NodeDeviceDescriptorManager {
   std::unique_ptr<Impl> impl_;
 };
 
-}  // namespace device
+}  // namespace hardware
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_DEVICE_NODE_DEVICE_DESCRIPTOR_MANAGER_H_
+#endif  // ONEFLOW_CORE_HARDWARE_NODE_DEVICE_DESCRIPTOR_MANAGER_H_
