+++ oneflow/core/job/env_global_objects_scope.cpp
@@ -34,7 +34,7 @@ limitations under the License.
 #include "oneflow/core/device/cudnn_conv_util.h"
 #include "oneflow/core/rpc/include/manager.h"
 #include "oneflow/core/transport/transport.h"
-#include "oneflow/core/device/node_device_descriptor_manager.h"
+#include "oneflow/core/hardware/node_device_descriptor_manager.h"
 #include "oneflow/core/vm/symbol_storage.h"
 #include "oneflow/core/framework/multi_client_session_context.h"
 #include "oneflow/core/framework/symbol_id_cache.h"
@@ -171,10 +171,10 @@ Maybe<void> EnvGlobalObjectsScope::Init(const EnvProto& env_proto) {
                                     GlobalProcessCtx::NumOfProcessPerNode());
   Global<ResourceDesc, ForSession>::New(GetDefaultResource(env_proto),
                                         GlobalProcessCtx::NumOfProcessPerNode());
-  Global<device::NodeDeviceDescriptorManager>::SetAllocated(
-      new device::NodeDeviceDescriptorManager());
+  Global<hardware::NodeDeviceDescriptorManager>::SetAllocated(
+      new hardware::NodeDeviceDescriptorManager());
   if (Global<ResourceDesc, ForEnv>::Get()->enable_debug_mode()) {
-    Global<device::NodeDeviceDescriptorManager>::Get()->DumpSummary("devices");
+    Global<hardware::NodeDeviceDescriptorManager>::Get()->DumpSummary("devices");
   }
   Global<ep::DeviceManagerRegistry>::New();
   Global<ThreadPool>::New(Global<ResourceDesc, ForSession>::Get()->ComputeThreadPoolSize());
@@ -249,7 +249,7 @@ EnvGlobalObjectsScope::~EnvGlobalObjectsScope() {
     Global<ResourceDesc, ForSession>::Delete();
   }
   Global<ResourceDesc, ForEnv>::Delete();
-  Global<device::NodeDeviceDescriptorManager>::Delete();
+  Global<hardware::NodeDeviceDescriptorManager>::Delete();
   CHECK_NOTNULL(Global<CtrlClient>::Get());
   CHECK_NOTNULL(Global<EnvDesc>::Get());
   Global<RpcManager>::Delete();
