+++ oneflow/core/job/session_global_objects_scope.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 #include "oneflow/core/job/session_global_objects_scope.h"
 #include "oneflow/core/control/ctrl_server.h"
 #include "oneflow/core/control/global_process_ctx.h"
-#include "oneflow/core/device/node_device_descriptor_manager.h"
+#include "oneflow/core/hardware/node_device_descriptor_manager.h"
 #include "oneflow/core/framework/load_library.h"
 #include "oneflow/core/job/available_memory_desc.pb.h"
 #include "oneflow/core/job/collective_boxing/scheduler.h"
@@ -41,7 +41,7 @@ limitations under the License.
 #include "oneflow/core/graph/task_stream_index_manager.h"
 
 #ifdef WITH_CUDA
-#include "oneflow/core/device/cuda_device_descriptor.h"
+#include "oneflow/core/hardware/cuda_device_descriptor.h"
 #endif  // WITH_CUDA
 
 namespace oneflow {
@@ -51,17 +51,17 @@ namespace {
 AvailableMemDescOfMachine GetAvailableMemDescOfMachine(int64_t rank) {
   AvailableMemDescOfMachine machine_mem_desc;
   const auto node_desc =
-      Global<device::NodeDeviceDescriptorManager>::Get()->GetNodeDeviceDescriptor(rank);
+      Global<hardware::NodeDeviceDescriptorManager>::Get()->GetNodeDeviceDescriptor(rank);
 #ifdef WITH_CUDA
   const auto cuda_device_list =
-      node_desc->GetDeviceDescriptorList(device::kCudaDeviceDescriptorClassName);
+      node_desc->GetDeviceDescriptorList(hardware::kCudaDeviceDescriptorClassName);
   CHECK(cuda_device_list);
   FOR_RANGE(int, i, 0, (Global<ResourceDesc, ForSession>::Get()->GpuDeviceNum())) {
     if (i >= cuda_device_list->DeviceCount()) {
       LOG(WARNING) << "Invalid CUDA device ordinal: rank " << rank << " ordinal " << i;
       machine_mem_desc.add_zone_size(0);
     } else {
-      const auto cuda_device = std::dynamic_pointer_cast<const device::CudaDeviceDescriptor>(
+      const auto cuda_device = std::dynamic_pointer_cast<const hardware::CudaDeviceDescriptor>(
           cuda_device_list->GetDevice(i));
       CHECK(cuda_device);
       machine_mem_desc.add_zone_size(cuda_device->GlobalMemorySizeBytes());
